requirements.txt
paho_mqtt==1.4.0
picamera==1.13
qiniu==7.2.2
